+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - [feat: support log path configuration parameters,](https://github.com/Tencent/spring-cloud-tencent/pull/1142)
 - [feat:add swagger report switch.](https://github.com/Tencent/spring-cloud-tencent/pull/1150)
 - [fix: dynamic routing using cookies.](https://github.com/Tencent/spring-cloud-tencent/pull/1153)
+- [fix:fix retry loadbalancer not working bug.](https://github.com/Tencent/spring-cloud-tencent/pull/1155)